Update ERCF_Software to preserve pressure advance during endless spool mode
Spool changes drop the pressure advance values to 0, This causes the pressure advance during endless spool mode to go to 0 but it does not ever reset due this being an unplanned change and thus no PA value is issued in gcode again. This saves the PA value before change and applies it again after the change.